# apps/backend/openapi.yaml
openapi: 3.1.0
info:
  title: AI Voice Assistant & Calendar Manager API
  description: |
    FastAPI backend for the AI-powered voice assistant and calendar manager.
    
    ## Features
    - Voice processing and natural language understanding
    - Calendar management and event scheduling
    - AI integration with OpenAI GPT-4 and Anthropic Claude
    - Real-time WebSocket communication
    - User authentication and authorization
    
    ## Authentication
    This API uses JWT Bearer token authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```
    
    ## Rate Limiting
    - 100 requests per minute for authenticated users
    - 10 requests per minute for unauthenticated users
    
  version: 1.0.0
  contact:
    name: AI Voice Assistant Team
    email: support@aivoiceassistant.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.aivoiceassistant.com
    description: Production server
security:
  - BearerAuth: []
paths:
...
  /api/v1/voice/process:
    post:
      summary: Process Voice Command
      description: Process voice input and return AI-generated response
      tags:
        - Voice
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VoiceProcessingRequest'
      responses:
        '200':
          description: Voice command processed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VoiceProcessingResponse'
        '400':
          description: Invalid voice data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/voice/history:
    get:
      summary: Get Voice Command History
      description: Retrieve user's voice command history
      tags:
        - Voice
      security:
        - BearerAuth: []
      parameters:
        - name: limit
          in: query
          description: Number of commands to retrieve
          required: false
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 100
        - name: offset
          in: query
          description: Number of commands to skip
          required: false
          schema:
            type: integer
            default: 0
            minimum: 0
      responses:
        '200':
          description: Voice command history retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/VoiceCommand'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    VoiceProcessingRequest:
      type: object
      required:
        - audio_data
        - user_id
      properties:
        audio_data:
          type: string
          description: Base64 encoded audio data
          example: "UklGRnoGAABXQVZFZm10IBAAAAABAAEAQB8AAEAfAAABAAgAZGF0YQoGAACBhYqFbF1fdJivrJBhNjVgodDbq2EcBj+a2/LDciUFLIHO8tiJNwgZaLvt559NEAxQp+PwtmMcBjiR1/LMeSwFJHfH8N2QQAoUXrTp66hVFApGn+DyvmwhBSuBzvLZiTYIG2m98OScTgwOUarm7blmGgU7k9n1unEiBC13yO/eizEIHWq+8+OWT..."
        user_id:
          type: string
          format: uuid
          description: User's unique identifier
          example: "123e4567-e89b-12d3-a456-426614174000"
        session_id:
          type: string
          description: Voice session identifier
          example: "session_123"
    VoiceProcessingResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            text:
              type: string
              description: Transcribed text from audio
              example: "Schedule a meeting with the team for tomorrow at 2pm"
            intent:
              type: string
              description: Detected intent from the voice command
              example: "schedule_meeting"
            confidence:
              type: number
              format: float
              minimum: 0
              maximum: 1
              description: Confidence score for the transcription and intent
              example: 0.92
            entities:
              type: object
              description: Extracted entities from the voice command
              example:
                attendees: ["team"]
                time: "tomorrow at 2pm"
                duration: "1 hour"
            suggested_actions:
              type: array
              items:
                type: string
              description: Suggested actions based on the voice command
              example: ["Create calendar event", "Send invitations", "Set reminder"]
    VoiceCommand:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Voice command's unique identifier
          example: "123e4567-e89b-12d3-a456-426614174000"
        user_id:
          type: string
          format: uuid
          description: User's unique identifier
          example: "123e4567-e89b-12d3-a456-426614174000"
        command:
          type: string
          description: Original voice command text
          example: "Schedule a meeting with the team for tomorrow at 2pm"
        intent:
          type: string
          description: Detected intent
          example: "schedule_meeting"
        confidence:
          type: number
          format: float
          minimum: 0
          maximum: 1
          description: Confidence score
          example: 0.92
        entities:
          type: array
          items:
            type: object
            properties:
              type:
                type: string
                description: Entity type
                example: "attendee"
              value:
                type: string
                description: Entity value
                example: "team"
              confidence:
                type: number
                format: float
                description: Entity confidence score
                example: 0.9
        created_at:
          type: string
          format: date-time
          description: Command creation timestamp
          example: "2024-01-15T08:30:00Z"
...
    Pagination:
      type: object
      properties:
        total:
          type: integer
          description: Total number of items
          example: 100
        limit:
          type: integer
          description: Number of items per page
          example: 20
        offset:
          type: integer
          description: Number of items skipped
          example: 0
        has_more:
          type: boolean
          description: Whether there are more items available
          example: true
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              description: Error code
              example: "validation_error"
            message:
              type: string
              description: Error message
              example: "Invalid input data"
            details:
              type: object
              description: Additional error details
              example:
                field: "email"
                issue: "Invalid email format"
...